+++ utils/frame/benchmarking-cli/src/storage/weights.hbs
@@ -17,8 +17,9 @@
 
 /// Storage DB weights for the `{{runtime_name}}` runtime and `{{db_name}}`.
 pub mod constants {
+	use frame_support::weights::constants;
 	use sp_core::parameter_types;
-	use sp_weights::{constants, RuntimeDbWeight};
+	use sp_weights::RuntimeDbWeight;
 
 	parameter_types! {
 		{{#if (eq db_name "ParityDb")}}
